fix: dead except clause, duplicate _resolve_coroutine, nested body regex
Three fixes identified during code review of PR #25:

1. BUG: Dead except OSError clause in simulation.py
   - In _inject_object_into_scene(), 'except OSError' after 'except Exception'
     is unreachable since Exception is a parent class of OSError.
   - Removed the dead clause.

2. DESIGN: Duplicate _resolve_coroutine() definition
   - Both robot.py and simulation.py defined identical _resolve_coroutine()
     and robot.py also had _run_async() — same logic, copy-pasted.
   - Extracted to strands_robots/_async_utils.py as the single source of truth.
   - Both modules now import from the shared utility.

3. BUG: Regex edge case in _eject_body_from_scene()
   - The non-greedy regex .*? with re.DOTALL matched the FIRST </body>,
     which for nested <body> elements would be a child's closing tag,
     leaving the parent body partially intact (orphaned tags, broken XML).
   - Replaced with a nesting-aware parser that tracks depth to find
     the correct matching </body> for the target element.

All 71 existing tests pass.
